+++ src/models/mod.rs
@@ -97,7 +97,7 @@ pub enum Element {
 impl Element {
     #[allow(unused)]
     /// Returns the constant triples of the element, if applicable.
-    pub(crate) fn get_triples(&self, x_vec: &Vec<f64>) -> Option<Triples<Numeric, 4>> {
+    pub(crate) fn get_triples(&self, x_vec: &Vec<Numeric>) -> Option<Triples<Numeric, 4>> {
         match self {
             Element::VSource(ele) => Some(ele.triples()),
             Element::Resistor(ele) => Some(ele.triples()),
@@ -111,7 +111,7 @@ impl Element {
 
     #[allow(unused)]
     /// Returns the time variant pairs of the element, if applicable.
-    pub(crate) fn get_pairs(&self, x_vec: &Vec<f64>) -> Option<Pairs<Numeric, 2>> {
+    pub(crate) fn get_pairs(&self, x_vec: &Vec<Numeric>) -> Option<Pairs<Numeric, 2>> {
         match self {
             Element::Diode(ele) => Some(ele.pairs(x_vec)),
             Element::Mos0(ele) => Some(ele.pairs(x_vec)),
@@ -155,7 +155,7 @@ impl Element {
     }
 
     /// Returns the nonlinear triples. Nonlinear Triples are Dependend on Vector x
-    pub(crate) fn get_nonlinear_triples(&self, x_vec: &Vec<f64>) -> Option<Triples<Numeric, 4>> {
+    pub(crate) fn get_nonlinear_triples(&self, x_vec: &Vec<Numeric>) -> Option<Triples<Numeric, 4>> {
         match self {
             Element::Diode(ele) => Some(ele.triples(x_vec)),
             Element::Mos0(ele) => Some(ele.triples(x_vec)),
@@ -164,7 +164,7 @@ impl Element {
     }
 
     /// Returns the nonlinear pairs of the element, if applicable.
-    pub(crate) fn get_nonlinear_pairs(&self, x_vec: &Vec<f64>) -> Option<Pairs<Numeric, 2>> {
+    pub(crate) fn get_nonlinear_pairs(&self, x_vec: &Vec<Numeric>) -> Option<Pairs<Numeric, 2>> {
         match self {
             Element::Diode(ele) => Some(ele.pairs(x_vec)),
             Element::Mos0(ele) => Some(ele.pairs(x_vec)),
@@ -182,7 +182,7 @@ impl Element {
     }
 
     /// Returns the ac triples. Ac Triples are dependend on f
-    pub(crate) fn get_ac_triples(&self, freq: f64) -> Option<Triples<ComplexNumeric, 4>> {
+    pub(crate) fn get_ac_triples(&self, freq: Numeric) -> Option<Triples<ComplexNumeric, 4>> {
         match self {
             Element::Diode(_) => None,
             Element::Mos0(_) => None,
@@ -195,7 +195,7 @@ impl Element {
     }
 
     /// Returns the ac pairs of the element, if applicable.
-    pub(crate) fn get_ac_pairs(&self, _freq: f64) -> Option<Pairs<ComplexNumeric, 2>> {
+    pub(crate) fn get_ac_pairs(&self, _freq: Numeric) -> Option<Pairs<ComplexNumeric, 2>> {
         match self {
             Element::Diode(_) => None,
             Element::Mos0(_) => None,
